lib: Change rdrbuff API to use correct types
The head and tail alloc and release operations were taking an int to
identify the idx instead of a ssize_t. The size was a ssize_t instead
of a size_t.

+++ include/ouroboros/shm_rdrbuff.h
@@ -67,18 +67,18 @@ int       shm_rdrbuff_remove(struct shm_rdrbuff  * rdrb,
 
 
 uint8_t * shm_du_buff_head_alloc(struct shm_rdrbuff * rdrb,
-                                 int                  idx,
-                                 ssize_t              size);
+                                 ssize_t              idx,
+                                 size_t               size);
 
 uint8_t * shm_du_buff_tail_alloc(struct shm_rdrbuff * rdrb,
-                                 int                  idx,
-                                 ssize_t              size);
+                                 ssize_t              idx,
+                                 size_t               size);
 
 int       shm_du_buff_head_release(struct shm_rdrbuff * rdrb,
-                                   int                  idx,
-                                   ssize_t              size);
+                                   ssize_t              idx,
+                                   size_t               size);
 
 int       shm_du_buff_tail_release(struct shm_rdrbuff * rdrb,
-                                   int                  idx,
-                                   ssize_t              size);
+                                   ssize_t              idx,
+                                   size_t               size);
 #endif /* OUROBOROS_SHM_RDRBUFF_H */
